requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
crispy-bootstrap5==0.7
cryptography==41.0.4
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.5
django-allauth==0.57.0
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
docutils==0.20.1
gunicorn==21.2.0
idna==3.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-http-client==3.3.7
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rollbar==1.0.0
sendgrid==6.10.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.9.0
urllib3==2.0.6
whitenoise==6.6.0
